refactor: centralize choice selector creation
